advanced/open-clip/requirements.txt
open_clip_torch
Pillow
torch
transformers
validators
